[refactor] Solar 엔티티 및 패키지명/필드명 변경
