+++ src/main/java/org/farmsystem/homepage/domain/minigame/solar/dto/SolarDTO.java
@@ -1,13 +1,13 @@
-package org.farmsystem.homepage.domain.minigame.solarstation.dto;
+package org.farmsystem.homepage.domain.minigame.solar.dto;
 
-import org.farmsystem.homepage.domain.minigame.solarstation.entity.SolarPowerStation;
+import org.farmsystem.homepage.domain.minigame.solar.entity.Solar;
 
 import java.time.LocalDateTime;
 
 public record SolarDTO( // TODO: Solar 무슨 DTO인지 (Response, Request) 써주기!
         LocalDateTime chargeStartTime
 ) {
-    public static SolarDTO from(SolarPowerStation station) {
+    public static SolarDTO from(Solar station) {
         return new SolarDTO(
                 station.getChargeStartedAt()
         );
