+++ src/main/java/org/farmsystem/homepage/domain/minigame/solar/service/SolarService.java
@@ -1,12 +1,12 @@
-package org.farmsystem.homepage.domain.minigame.solarstation.service;
+package org.farmsystem.homepage.domain.minigame.solar.service;
 
 
 import lombok.RequiredArgsConstructor;
 import org.farmsystem.homepage.domain.minigame.player.entity.Player;
 import org.farmsystem.homepage.domain.minigame.player.repository.PlayerRepository;
-import org.farmsystem.homepage.domain.minigame.solarstation.dto.SolarDTO;
-import org.farmsystem.homepage.domain.minigame.solarstation.entity.SolarPowerStation;
-import org.farmsystem.homepage.domain.minigame.solarstation.repository.SolarRepository;
+import org.farmsystem.homepage.domain.minigame.solar.dto.SolarDTO;
+import org.farmsystem.homepage.domain.minigame.solar.entity.Solar;
+import org.farmsystem.homepage.domain.minigame.solar.repository.SolarRepository;
 import org.farmsystem.homepage.global.error.exception.BusinessException;
 import org.farmsystem.homepage.global.error.ErrorCode;
 import org.springframework.stereotype.Service;
@@ -26,16 +26,16 @@ public class SolarService {
     }
 
     // 플레이어에게 발전소가 없으면 생성
-    private SolarPowerStation getOrCreateStation(Player player) {
+    private Solar getOrCreateStation(Player player) {
         return solarRepository.findByPlayer(player)
-                .orElseGet(() -> solarRepository.save(SolarPowerStation.createSolarStation(player)));
+                .orElseGet(() -> solarRepository.save(Solar.createSolarStation(player)));
     }
 
     // 플레이어의 태양광 발전소 조회
     @Transactional
     public SolarDTO getSolarStation(Long userId) {
         Player player = findPlayerOrThrow(userId);
-        SolarPowerStation station = getOrCreateStation(player);
+        Solar station = getOrCreateStation(player);
         return SolarDTO.from(station);
     }
 
@@ -43,7 +43,7 @@ public class SolarService {
     @Transactional
     public SolarDTO updateChargeTime(Long userId, SolarDTO request) {
         Player player = findPlayerOrThrow(userId);
-        SolarPowerStation station = getOrCreateStation(player);
+        Solar station = getOrCreateStation(player);
 
         station.updateChargeStartTime(request.chargeStartTime());
